+++ CHANGELOG.md
@@ -8,6 +8,7 @@ yyyy/mm/dd Version x.y.z
 - Improved CLI help text
 - Add `--for SECONDS` option to `crystalfontz listen` and `crystalfontz effects` that closes the commands after a certain amount of time
 - Renamed `client` contextmanager to `connection`
+- Renamed `TemperatureReport`'s `idx` attribute to `index`
 
 2025/01/06 Version 2.0.0
 ------------------------